test: move tla upstream tests to a separate file
